fix: default reporter from platform
