+++ pkg/commands/entrypoints/entrypoints.go
@@ -27,7 +27,6 @@ import (
 	"slices"
 	"sort"
 
-	"github.com/posener/complete/v2"
 	"golang.org/x/exp/maps"
 
 	"github.com/abcxyz/guardian/pkg/flags"
@@ -60,7 +59,6 @@ type EntrypointsCommand struct {
 
 	flags.CommonFlags
 
-	flagReporter                string
 	flagDestRef                 string
 	flagSourceRef               string
 	flagDetectChanges           bool
@@ -93,17 +91,6 @@ func (c *EntrypointsCommand) Flags() *cli.FlagSet {
 	c.CommonFlags.Register(set)
 
 	f := set.NewSection("COMMAND OPTIONS")
-
-	f.StringVar(&cli.StringVar{
-		Name:    "reporter",
-		Target:  &c.flagReporter,
-		Default: reporter.TypeNone,
-		Example: "github",
-		Usage:   fmt.Sprintf("The reporting strategy for Guardian status updates. Valid values are %q.", reporter.SortedReporterTypes),
-		Predict: complete.PredictFunc(func(prefix string) []string {
-			return reporter.SortedReporterTypes
-		}),
-	})
 
 	f.StringVar(&cli.StringVar{
 		Name:    "dest-ref",
@@ -188,7 +175,7 @@ func (c *EntrypointsCommand) Run(ctx context.Context, args []string) error {
 	}
 	c.platformClient = platform
 
-	rc, err := reporter.NewReporter(ctx, c.flagReporter, &reporter.Config{GitHub: c.platformConfig.GitHub})
+	rc, err := reporter.NewReporter(ctx, c.platformConfig.Reporter, &reporter.Config{GitHub: c.platformConfig.GitHub})
 	if err != nil {
 		return fmt.Errorf("failed to create reporter client: %w", err)
 	}
